Fixed "pnpm run dev" not working as intended
"type":"module" is needed for the vite/dev server to work with
the current setup. Checking
https://docs.npmjs.com/cli/v10/configuring-npm/package-json
there is no mention of "type":"module" so I initially deleted it from
the default setup since I did know what it was and no mention on the
website.

But it is mentioned here:
https://nodejs.org/docs/latest-v13.x/api/esm.html#esm_enabling

Has to do with using the new ES modules import statements and not the
older style of "require" import syntax.

Source:
https://stackoverflow.com/questions/61401475/why-is-type-module-in-package-json-file